google-photo-pictures-uploader/requirements.txt
requests
python-dotenv
google-auth-oauthlib